Data Science Salaries
So the data space is quite broad, different people with different roles and job descriptions and titles. And as one that's just fully entering the data space, I thought it'd be necessary for me to see how the payments are here, so I did my research and created a dashboard to analyze data experts payments by experience, job location, company size etc..

From my analysis, medium scale companies have had the most money invested in data experts in the past 4yrs, 87% of the total earnings of data experts come from medium scale companies, perhaps because there are a lot more SME's than large scale companies, and small scale companies don't have much need or can't afford to invest too much in data experts, because well, not that much data yet 🤷🏾‍♂️
What's a medium scale company? A medium scale company is one with a staff strength of 50-199 and a gross turnover of 25-100m, And they seem to need us the most.

I think it's interesting to also note the amazing growth in salaries over the years! Although, this trend is only constant with Medium scale companies. In large and small scale companies, there was fall in salaries from 2022.
And we see also the top 10 countries that are investing in data experts. A total of about half a billion dollars yearly to these 10 roles.